Issue #421 - Fixed sleep timeout going negative

+++ astyanax-core/src/main/java/com/netflix/astyanax/retry/BoundedExponentialBackoff.java
@@ -34,7 +34,7 @@ public class BoundedExponentialBackoff extends ExponentialBackoff {
 
     private final int maxSleepTimeMs;
 
-    public BoundedExponentialBackoff(int baseSleepTimeMs, int maxSleepTimeMs, int max) {
+    public BoundedExponentialBackoff(long baseSleepTimeMs, int maxSleepTimeMs, int max) {
         super(baseSleepTimeMs, max);
         this.maxSleepTimeMs = maxSleepTimeMs;
     }
