+++ astyanax-core/src/main/java/com/netflix/astyanax/retry/ExponentialBackoff.java
@@ -30,9 +30,9 @@ public class ExponentialBackoff extends SleepingRetryPolicy {
     private final int MAX_SHIFT = 30;
     
     private final Random random = new Random();
-    private final int baseSleepTimeMs;
+    private final long baseSleepTimeMs;
 
-    public ExponentialBackoff(int baseSleepTimeMs, int maxAttempts) {
+    public ExponentialBackoff(long baseSleepTimeMs, int maxAttempts) {
         super(maxAttempts);
         this.baseSleepTimeMs = baseSleepTimeMs;
     }
@@ -46,7 +46,7 @@ public class ExponentialBackoff extends SleepingRetryPolicy {
         return baseSleepTimeMs * Math.max(1, random.nextInt(1 << attempt));
     }
     
-    public int getBaseSleepTimeMs() {
+    public long getBaseSleepTimeMs() {
         return baseSleepTimeMs;
     }
